Remove several redundant update() calls in GUI.
When idle, the GUI is using up to 30% CPU to update its status very
often.  It turns out that several of the widgets were getting the
update() call when their values changed.  This was thought to be safe,
but turns out to be too manu GUI updates.

Now, the GUI is totally disconnected from RT events in the audio
thread.  The only way the GUI is getting updated is by the timer that
is set up in PlayerWidget::_setup_signals_and_slots()... which is
currently set to update every 200 ms (5 times/sec).  This results in
about 3-5% CPU.

However, if compositing is enabled, the X server may still be using
20-30% CPU to calculate the compositing features.