PlayerWidget: Update the cursor on mousePressEvent().
When it comes to mouse/cursor interaction, several decisions are made
based on what the current cursor type is.  The assuption is that a
mouseMoveEvent() will always happen before a mousePressEvent().  In a
traditional desktop/mouse environment... this usually holds true.

However, in a touchscreen environment you only get a mouseMoveEvent()
when "clicked."  Thus, you almost always have the wrong cursor and the
window does the wrong thing.  It's not uncommon to get stuck in a
resizing mode.

This commit updates the cursor on every mousePressEvent().