Engine: Handle more segment sizes, and adjust if it changes.
If the segment size was large (e.g. 4096) then playing would stall.
In addition, large segment sizes were not being handled at all because
of fixed buffer sizes.  This improves some of that.

However, it appears that the first segment or so of input audio is
being truncated.  This is very apparent for the 4096 size
segment... but not so much when <= 512.